+++ react/__tests__/SearchResult.test.js
@@ -8,7 +8,6 @@ import SearchResult from '../SearchResult'
 describe('<SearchResult /> component', () => {
   let wrapper = null
   let searchQueryMock = null
-  let facetsQueryMock = null
 
   beforeEach(() => {
     const messages = require('../locales/en-US')
@@ -19,11 +18,12 @@ describe('<SearchResult /> component', () => {
         map: 'c',
         orderBy: 'OrderByTopSaleDESC',
       },
+      facets: {
+        Departments: [],
+        Brands: [],
+        PriceRanges: [],
+      },
       products: [],
-    }
-
-    facetsQueryMock = {
-      loading: false,
     }
 
     wrapper = mount(
@@ -34,8 +34,8 @@ describe('<SearchResult /> component', () => {
           map="c"
           path="eletronics/s"
           page={1}
+          orderBy="OrderByTopSaleDESC"
           searchQuery={searchQueryMock}
-          facetsQuery={facetsQueryMock}
         />
       </IntlProvider>
     )