+++ react/components/SelectedFilters.js
@@ -1,5 +1,4 @@
 import PropTypes from 'prop-types'
-import { concat, keys, map, reduce } from 'ramda'
 import React, { Component } from 'react'
 import { intlShape, injectIntl } from 'react-intl'
 import { Link } from 'render'
@@ -12,66 +11,46 @@ import FiltersContainer from './FiltersContainer'
  */
 class SelectedFilters extends Component {
   static propTypes = {
-    /** If the filter is collapsed or not. */
-    opened: PropTypes.bool,
     /** Selected filters. */
-    selecteds: PropTypes.shape({
-      Departments: PropTypes.arrayOf(PropTypes.string),
-      Brands: PropTypes.arrayOf(PropTypes.string),
-      SpecificationFilters: PropTypes.arrayOf(PropTypes.string),
-      FullText: PropTypes.arrayOf(PropTypes.string),
-    }).isRequired,
+    selecteds: PropTypes.arrayOf(PropTypes.shape({
+      Name: PropTypes.string,
+      Link: PropTypes.string,
+      slug: PropTypes.string,
+      type: PropTypes.string,
+    })).isRequired,
     getLinkProps: PropTypes.func,
     intl: intlShape,
   }
 
   static defaultProps = {
-    opened: true,
     selecteds: [],
-  }
-
-  getSelectedFilters() {
-    const selecteds = this.props.selecteds
-    return reduce(
-      (accumulator, filterName) =>
-        concat(
-          accumulator,
-          map(term => {
-            return { key: `${filterName}:${term}`, value: term }
-          }, selecteds[filterName])
-        ),
-      [],
-      keys(selecteds)
-    )
   }
 
   render() {
     const { intl } = this.props
-    const selectedFilters = this.getSelectedFilters()
-
     const title = intl.formatMessage({ id: 'search.selected-filters' })
-
     return (
       <FiltersContainer
         title={title}
-        filters={selectedFilters}
+        filters={this.props.selecteds}
       >
-        {filter => {
+        {({ Name: name, Link: link, type, slug }) => {
           const pagesArgs = this.props.getLinkProps({
-            opt: { Name: filter.value },
+            name,
+            link,
+            type,
             isSelected: true,
           })
-
           return (
             <Link
-              key={filter.key}
+              key={slug}
               className="w-100 flex clear-link"
               page={pagesArgs.page}
               params={pagesArgs.params}
               query={pagesArgs.queryString}>
               <Checkbox
                 checked
-                label={filter.value}
+                label={name}
                 name="default-checkbox-group"
                 value=""
                 onChange={evt => {